start breaking page.spec.js into smaller files
